+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added support for the Arduino Zero and Due boards (and related)
 - Added method to LEDImage for drawing circles
+- Added a Red-Blue-Green bit layout mode for the RGB matrix object
 
 ## [1.0.1] - 2017-12-24
 ### Changed